+++ src/lib/events.ts
@@ -10,6 +10,7 @@ import { HooksContext } from '../client/failure/event-adapter';
 import { MapInterpreterEventAdapter } from '../client/failure/map-interpreter-adapter';
 import { FailurePolicyReason } from '../client/failure/policy';
 import { UnexpectedError } from '../internal/errors';
+import { HttpResponse, RequestParameters } from '../internal/interpreter/http';
 import { FetchInstance } from '../internal/interpreter/http/interfaces';
 
 const debug = createDebug('superface:events');
@@ -112,6 +113,11 @@ export type FailureContext = EventContextBase & {
   profile: string;
   usecase: string;
   provider: string;
+};
+
+export type AuthenticateContext = EventContextBase & {
+  resourceRequest?: RequestParameters;
+  previousResponse?: HttpResponse;
 };
 
 type VoidEventTypes = {